+++ source/ambient.d.ts
@@ -18,6 +18,10 @@ declare module "*/package.json"
 declare module "require-from-string"
 declare module "node-eval"
 declare module "node-cleanup"
+declare module "cli-interact"
+
+declare module "hyperlinker"
+declare module "supports-hyperlinks"
 
 // declare module "require-from-string" {
 //   export interface RequireOptions {