+++ source/dsl/GitHubDSL.ts
@@ -234,6 +234,10 @@ export interface GitHubUser {
    * Whether the user is an org, or a user
    */
   type: "User" | "Organization"
+  /**
+   * The url for a users's image
+   */
+  avatar_url: string
 }
 
 /**