+++ CHANGELOG.md
@@ -3,8 +3,7 @@ developers, so please limit technical // terminology to in here.
 
 // ### Master
 
-### 2.0.4
-
+* Adds a new command for getting set up: `danger init` - [@orta][]
 * Fix double negative in documentation. [@dfalling][]
 * Fix `gloabally` typo in documentation. [@dfalling][]
 
@@ -740,3 +739,4 @@ Not usable for others, only stubs of classes etc. - [@orta][]
 [@orta]: https://github.com/orta
 [@ashfurrow]: https://github.com/ashfurrow
 [@joarwilk]: https://github.com/joarwilk
+[@dfalling]: https://github.com/dfalling